// src/api/generated/model/countryCreate.ts
/**
 * Generated by orval v6.27.1 🍺
 * Do not edit manually.
 * API Geo
 * OpenAPI spec version: 0.1.0
 */
import type { CountryCreateNeedAutomaticUpdate } from "./countryCreateNeedAutomaticUpdate";

export interface CountryCreate {
  iso3116_alpha2: string;
  iso3166_alpha3: string;
  latitude: number;
  longitude: number;
  name: string;
  need_automatic_update?: CountryCreateNeedAutomaticUpdate;
  osm_id: string;
  osm_type: string;
  phone_code: string;
  phone_mask: string;
}